moved recentArticles into HomePage.jsx and deleted RecentArticle component
