+++ src/components/ArticleCard/ArticleCard.jsx
@@ -14,4 +14,4 @@ const ArticleCard = ({ article }) => {
     )
 }
 
-export default ArticleCard
+export default ArticleCard
